move searcher test
